+++ src/app/api/tendency/[id]/route.ts
@@ -1,23 +1,20 @@
-import { NextRequest, NextResponse } from 'next/server';
+import { NextRequest, NextResponse } from "next/server";
 
 // 테스트용 mock 데이터
 const tendencySamples: Record<string, number[]> = {
-  '3': [60, 90, 40, 70, 25, 10], // 둥근 그래프용
-  '4': [75, 85, 60, 55, 35, 20]  // 각진 그래프용
+  "3": [60, 90, 40, 70, 25, 10],
+  "4": [75, 85, 60, 55, 35, 20],
 };
 
-export async function GET(
-  req: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  const { id } = params;
-  const result = tendencySamples[id];
+export async function GET(req: NextRequest) {
+  const url = new URL(req.url);
+  const id = url.pathname.split("/").pop(); // 또는 regex로 추출
 
-  if (!result) {
+  if (!id || !tendencySamples[id]) {
     return NextResponse.json(
       {
         success: false,
-        message: `해당 id(${id})에 대한 통신 성향 데이터가 없습니다.`
+        message: `해당 id(${id})에 대한 통신 성향 데이터가 없습니다.`,
       },
       { status: 404 }
     );
@@ -25,6 +22,6 @@ export async function GET(
 
   return NextResponse.json({
     success: true,
-    data: result
+    data: tendencySamples[id],
   });
 }
